Projects/Leads-green/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1

# Supabase
supabase==2.9.1

# HTTP
httpx==0.27.2
requests==2.32.3

# Scraping
playwright==1.47.0
beautifulsoup4==4.12.3
lxml==5.3.0
praw==7.7.1  # Reddit

# AI
openai==1.54.0

# Utilities
python-dateutil==2.9.0.post0
pytz==2024.2
phonenumbers==8.13.45
tenacity==9.0.0  # retries
loguru==0.7.2

# Tools/scripts
tabulate==0.9.0
click==8.1.7
